requirements.txt
streamlit
torch
torchvision
fastai
scikit-image
Pillow
